Added fifty-fifty button & a coin meter
I added the fifty-fifty button which eliminates two wrong answers. The
button is only enable if the player has three coins, else the button
will be disable.

I also added a coin meter, which shows how many coins the users has, if
the user answers wrong a question a coin will be substracted, if the
user answers the question right a coin will be added. Users will start
will 3 coins.